Lexical: Extracted & merged heading & quote nodes
