+++ resources/js/wysiwyg/lexical/core/__tests__/unit/LexicalSerialization.test.ts
@@ -8,11 +8,12 @@
 
 import {$createLinkNode} from '@lexical/link';
 import {$createListItemNode, $createListNode} from '@lexical/list';
-import {$createHeadingNode, $createQuoteNode} from '@lexical/rich-text';
 import {$createTableNodeWithDimensions} from '@lexical/table';
 import {$createParagraphNode, $createTextNode, $getRoot} from 'lexical';
 
 import {initializeUnitTest} from '../utils';
+import {$createHeadingNode} from "@lexical/rich-text/LexicalHeadingNode";
+import {$createQuoteNode} from "@lexical/rich-text/LexicalQuoteNode";
 
 function $createEditorContent() {
   const root = $getRoot();
